manim-service/requirements.txt
manim==0.17.3
boto3==1.28.38
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
pydantic==2.3.0
python-dotenv==1.0.0
requests==2.31.0
